services/orders_service/pyproject.toml
[project]
name = "orders-service"
version = "0.1.0"
description = "A service for managing orders in an e-commerce application. See README for more details."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings >=2.10.1",
    "opentelemetry-distro>=0.56b0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "uvicorn[standard]>=0.35.0",
    "redis>=4.6.0",
    "aiokafka>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.12.0",
    "ruff>=0.12.3",
]


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
files = ["src"]
check_untyped_defs = true
